--- __init__.py
@@ -1,5 +1,0 @@
-import os
-import sys
-
-sys.path.append(os.path.dirname(os.path.abspath(__file__)))
-__all__ = ['config', 'helpers', 'main', 'contexts.py', 'mqtt_client', 'loaders']
