+++ no_index_requirements.txt
@@ -1,4 +1,3 @@
-wheel
 skabenproto>=1.10
 pygame>=1.9.4
 paho-mqtt>=1.4.0
