+++ requirements.txt
@@ -8,4 +8,5 @@ pyaml>=18.0.0
 filelock>=3.0.12,<3.1
 pytest
 flake8
+wheel
