+++ skabenclient/device.py
@@ -12,20 +12,26 @@ class BaseDevice:
 
     config_class = DeviceConfig
 
-    def __init__(self, system_config, device_config_path):
+    def __init__(self, system_config, device_config):
         if not isinstance(system_config, SystemConfig):
-            raise Exception(f'config object is not a SystemConfig, but {type(system_config)} instead')
-        # get only necessary from system config
+            raise Exception(f'system_config is not a SystemConfig, but {type(system_config)} instead')
+        if not isinstance(device_config, self.config_class):
+            raise Exception(f'device_config is not a {self.config_class}, but {type(device_config)} instead')
+        # assign system config
+        self.system = system_config
         self.q_int = system_config.get('q_int')
         self.uid = system_config.get('uid')
         self.logger = system_config.logger()
         # assign device ingame config
-        self.config = self.config_class(device_config_path)
+        self.config = device_config
         self.config.load()  # load and update current running conf
 
     def run(self):
         """ Abstract method for run device """
-        raise NotImplementedError(f"{self} is abstract and cannot be started")
+        print('application is starting...')
+        self.logger.debug(f'device starting as: {self.system} \n {self.config}')
+        event = make_event('device', 'reload')
+        self.q_int.put(event)
 
     def state_update(self, data):
         """ Update device configuration from user actions
