--- skabenclient/ts
@@ -1,1 +1,0 @@
-0
